+++ docs/docs-new/components/mdx/Grid/Grid.tsx
@@ -1,32 +1,31 @@
+import classnames from 'classnames/bind';
 import React from 'react';
+import styles from './Grid.module.css';
+
+const cn = classnames.bind(styles);
 
 export type GridProps = {
   children: React.ReactNode;
   cols?: number;
-  imageSize?: [ width?: number, height?: number];
 };
 
 export const GridContext = React.createContext('grid');
 
 const defaultProps = {
   cols: 3,
-  imageSize: [],
 };
 export const Grid = ({
   children,
   ...restProps
 }: GridProps) => {
   const normalizedProps = { ...defaultProps, ...restProps };
-  const settingsString = JSON.stringify(normalizedProps);
-  const className = `grid__col-${normalizedProps.cols}`;
+  const classNames = cn('Grid', `Grid--${normalizedProps.cols}`);
 
   return (
-    <GridContext.Provider value={settingsString}>
-      <div className="ant-row">
-        <div className={className}>
-          {children}
-        </div>
+    <div className="ant-row">
+      <div className={classNames}>
+        {children}
       </div>
-    </GridContext.Provider>
+    </div>
   );
 };