+++ docs/docs-new/components/mdx/index.ts
@@ -1,3 +1,4 @@
+import * as Buttons from '@/components/common/Button/Button';
 import { CodeTabs } from "@/components/mdx/CodeTabs";
 import { CubeQueryResultSet } from "@/components/mdx/CubeQueryResultSet";
 import {
@@ -14,10 +15,23 @@ import { InlineButton } from "@/components/mdx/InlineButton/InlineButton";
 import { LoomVideo } from "@/components/mdx/LoomVideo/LoomVideo";
 import { ParameterTable } from "@/components/mdx/ReferenceDocs/ParameterTable";
 import { Snippet, SnippetGroup } from "@/components/mdx/Snippets/SnippetGroup";
-import { Screenshot } from '@/components/mdx/Screenshot';
+import { Diagram, Screenshot } from '@/components/mdx/Screenshot';
 import { YouTubeVideo } from '@/components/mdx/YouTubeVideo/YouTubeVideo';
+import { CaseStudyPromoBlock } from '@/components/mdx/CaseStudyPromoBlock/CaseStudyPromoBlock';
+import { CloudPromoBlock } from '@/components/mdx/CloudPromoBlock/CloudPromoBlock';
+import { EventPromoBlock } from '@/components/mdx/EventPromoBlock/EventPromoBlock';
+import { H1 } from '@/components/overrides/Headings/H1';
+import { Link } from '../overrides/Anchor/Link';
+import { Table } from '@/components/overrides/Table/Table';
+import { Td } from '@/components/overrides/Table/Td';
+import { Th } from '@/components/overrides/Table/Th';
+import { Tr } from '@/components/overrides/Table/Tr';
 
 export const components = {
+  ...Buttons,
+  CaseStudyPromoBlock,
+  CloudPromoBlock,
+  EventPromoBlock,
   DangerBox,
   InfoBox,
   SuccessBox,
@@ -37,5 +51,14 @@ export const components = {
   CodeTabs,
   Btn: InlineButton,
   Screenshot,
-  YouTubeVideo
+  Diagram,
+  YouTubeVideo,
+
+  // Overrides
+  h1: H1,
+  a: Link,
+  table: Table,
+  td: Td,
+  th: Th,
+  tr: Tr,
 };