+++ docs/docs-new/pages/guides/recipes/analytics/_meta.js
@@ -1,0 +1,6 @@
+module.exports = {
+  "active-users": "Daily, Weekly, Monthly Active Users (DAU, WAU, MAU)",
+  "event-analytics": "Implementing event analytics",
+  "cohort-retention": "Implementing retention analysis & cohorts",
+  "funnels": "Implementing Funnel Analysis"
+}