+++ docs/docs-new/pages/product/api-sql/_meta.js
@@ -1,5 +1,0 @@
-module.exports = {
-  "overview": "Overview",
-  "security": "Authentication and Authorization",
-  "joins": "Joins"
-}