+++ docs/docs-new/pages/product/getting-started/_meta.js
@@ -1,5 +1,5 @@
 module.exports = {
-  "overview": "Overview",
   "core": "Cube Core",
-  "cloud": "Cube Cloud"
+  "cloud": "Cube Cloud",
+  "migrate-from-core": "Migrate from Cube Core"
 }